backend/requirements.txt
# Core
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.9.1

# Database
motor==3.7.0
pymongo==4.12.1

# Auth
PyJWT==2.10.1
bcrypt==4.3.0
python-multipart==0.0.20

# LLM
groq==0.25.0
openai==1.82.0

# RAG
qdrant-client==1.11.3
sentence-transformers==4.1.0
langchain==0.3.25
langchain-community==0.3.24
protobuf<6.0.0
pypdf==5.6.0
python-docx==1.1.2

# Multimodal
easyocr==1.7.2
openai-whisper==20250625
Pillow==11.2.1

# NLP
langdetect==1.0.9
textblob==0.18.0.post0
nltk==3.9.1

# Rate Limiting
slowapi==0.1.9

# Utilities
aiofiles==24.1.0
httpx==0.28.1
python-jose[cryptography]==3.5.0
